Mar. 25 - RPi Finalizations

+++ load_meta.py
@@ -1,16 +1,15 @@
 # !/usr/local/bin/python
-import subprocess
-import tensorflow as tf
 import os
-import numpy as np
 import glob
 import sys,argparse
-#import braille_util
+import braille_util
 import time
 start_time = time.time()
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
-
 print('>>>'+str(time.strftime('[%H:%M:%S]'))+' '+"Preloading Modules into memory")
+import subprocess
+import tensorflow as tf
+import numpy as np
 
 #PARAMS#
 num_mfccs = 13
@@ -47,7 +46,7 @@ out3 = graph.get_tensor_by_name("decoder/CTCGreedyDecoder:3")
 
 print("Waiting for audio...")
 
-def features(rawsnd,num):
+def features(rawsnd,num):    
     import librosa
     import librosa.feature as lib_feat
     x, sample_rate = librosa.load(rawsnd, sr=16000)
@@ -81,17 +80,18 @@ def process(f,i,s):
     print(os.path.getsize(f))
     print('>>>'+str(time.strftime('[%H:%M:%S]'))+' '+"Calculating Features")
     dat=features(str(f),num_mfccs)
-    feed_dict = {i: [dat],s: [len(dat)]}
+    feed_dict = {i:[dat],s:[len(dat)]}
     print('>>>'+str(time.strftime('[%H:%M:%S]'))+' '+"Running Outputs")
     z = sess.run((out,out1,out2,out3),feed_dict)
     d = decode_to_chars(z[1])
     print('>>>'+str(time.strftime('[%H:%M:%S]'))+' '+" > > > Output:"+str(d))
-    #seq = braille_util.seq2braille(d)
-    #braille_util.disp(seq,0.1)
+    seq2 = braille_util.seq2braille(d)
+    braille_util.disp(seq2,0.1)
     print("Execution time: %s seconds" % (time.time() - start_time))
 
 while True:
-    if 'check' in os.listdir('_dir'):
+    #print('prevstate',prevstate,'state',state)
+    if 'check' in os.listdir('_dir') and os.path.isfile('_dir/gpio_on'):
         process('_dir/tmp.wav',inp,seq)
         os.remove('_dir/check')
         os.remove('_dir/tmp.wav')
